Extract ModelPreparer class from prepare_model class to allow unit tests
